Rewritten plugin
- Added !tracers cmd
- Added tracers
- Added player trails
- Changed plugin structure (cause i couldn't find anything what i wrote ffs)
